+++ services/system/internal/logic/getmenutreelogic.go
@@ -35,13 +35,13 @@ func (l *GetMenuTreeLogic) GetMenuTree(in *system.Empty) (*system.SysMenuTreeRes
 			Id:        m.Id,
 			ParentId:  m.ParentId,
 			Name:      m.Name,
-			MenuType:  int32(m.MenuType),
+			MenuType:  m.MenuType,
 			Path:      m.Path,
 			Component: m.Component,
 			Icon:      m.Icon,
-			Sort:      int32(m.Sort),
-			Visible:   int32(m.Visible),
-			Status:    int32(m.Status),
+			Sort:      m.Sort,
+			Visible:   m.Visible,
+			Status:    m.Status,
 			Perms:     m.Perms,
 		}
 		data = append(data, item)
